Take out aor_flag, fix a fault in OpenJava
1. For the aor_flag:
The boolean flag aor_flag was originally set in AOIU and AODU. But the
reason of using it is still unclear. For this block of code:
c = a + b;
c = b;
return c;

if only choose AOIU, the current implementation gives two AOIU mutants:
c = a + b;
c = -b;   //
return c;

and
c = a + b;
c = b;
return -c;  //

However, if enabling AOIU and AORB together (the flag will be set to
true), we will get three AOIU mutants. The additional AOIU mutant is:
c = -a + b; //
c = b;
return c;

Four AORB will be:
c = a * b;
                c = a / b;
                c = a % b;
c = a - b;

So if we have AORB, we even get more AOIU. We decide to take out the
aor_flag before we can further identify a possible reason for it.

2. For the OpenJava:
The initializer statement of for-loop is stored as an array of
VariableDeclarator, but the chilrenAccept() method only looks for
ParseTree objects. This results in omitting parsing the initializer
statement in for-loop.
in “for(int i = a; i < 10; ++i)”, a will be parsed as FieldAccess type.

We updated OpenJava to make it correctly parse the for-loop.